03_few_shot_image_classifier/requirements.txt
gradio>=4.20.0
transformers>=4.35.0
torch>=2.0.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
